Segwit txn support

+++ src/EBTC.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.26;
 
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+import "openzeppelin-contracts/contracts/token/ERC20/ERC20.sol";
 
 contract EBTC is ERC20 {
     error BridgeOnly();
 
-    event TokenMinted(address to, uint amount);
+    event TokenMinted(address to, uint256 amount);
 
     address bridge;
